# api-gateway/cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This project generates random company details based on user input. It consists of an API Gateway, an Authentication Service, and a Company Service, each serving a specific purpose in the system.",
        "title": "Random Company Generator Microservices",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/sign-in": {
            "post": {
                "description": "API For User To Sign In",
                "tags": [
                    "Authentication"
                ],
                "summary": "User Sign In",
                "operationId": "User SignIn",
                "parameters": [
                    {
                        "description": "Sign In Details",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SignIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Sign In Completed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SignIn"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Inputs",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "User Not Exist With This Details",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/sign-up": {
            "post": {
                "description": "API For User To Sign Up",
                "tags": [
                    "Authentication"
                ],
                "summary": "User Sign Up",
                "operationId": "User SignUp",
                "parameters": [
                    {
                        "description": "Sign Up Details",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Sign Up Completed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SignUp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Inputs",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "User Already Exist",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/company": {
            "post": {
                "description": "API For User To Create Random Company Details",
                "tags": [
                    "Company"
                ],
                "summary": "Random Company Creating",
                "operationId": "Company Create",
                "parameters": [
                    {
                        "description": "Company Create Details",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CompanyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Company Details Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Company"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.CompanyRequest": {
            "type": "object",
            "required": [
                "ceo_name",
                "company_name",
                "total_employees"
            ],
            "properties": {
                "ceo_name": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 3
                },
                "company_name": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 3
                },
                "total_employees": {
                    "type": "integer",
                    "maximum": 5000,
                    "minimum": 100
                }
            }
        },
        "request.SignIn": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 6
                }
            }
        },
        "request.SignUp": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 6
                }
            }
        },
        "response.Company": {
            "type": "object",
            "properties": {
                "ceo": {
                    "$ref": "#/definitions/response.Employee"
                },
                "departments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Department"
                    }
                },
                "industry": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "totalEmployees": {
                    "type": "integer"
                }
            }
        },
        "response.Department": {
            "type": "object",
            "properties": {
                "leader": {
                    "$ref": "#/definitions/response.Employee"
                },
                "name": {
                    "type": "string"
                },
                "teams": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Team"
                    }
                },
                "totalEmployees": {
                    "type": "integer"
                }
            }
        },
        "response.Employee": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "hireDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "number"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.SignIn": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expire_at": {
                    "type": "string"
                }
            }
        },
        "response.SignUp": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "response.Team": {
            "type": "object",
            "properties": {
                "manager": {
                    "$ref": "#/definitions/response.Employee"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Employee"
                    }
                },
                "name": {
                    "type": "string"
                },
                "totalEmployees": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Add prefix of Bearer before token. For example, \"Bearer token\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}